Change $EXT_BUILD_DEPS dir to support multi-arch builds
This was set to `$EXT_BUILD_ROOT/bazel_foreign_cc_deps_" + lib_name`.
When building the same target (cmake, etc.) for multiple architectures, the
deps for each architecture try to link over each other, failing the build.

Make this dir relative to the target's dir in the exec root - then the target's
deps are constrained to its output dir